各機能をコンポーネントに分けダッシュボードとして表示
